modified:   src/App.tsx 	modified:   src/index.tsx 	renamed:    src/App.css -> src/style/App.css 	renamed:    src/general.css -> src/style/general.css 	renamed:    src/index.css -> src/style/index.css
