+++ src/index.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import "./index.css";
-import "./general.css";
+import "./style/index.css";
+import "./style/general.css";
 
 import App from "./App.tsx";
 import reportWebVitals from "./reportWebVitals.js";
